% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{analyseBranchPointTest}
\alias{analyseBranchPointTest}
\title{This tests the input to analyseBranchPoint.}
\usage{
analyseBranchPointTest(
  attributes,
  pseudotime,
  randomizationParams,
  statistic,
  start,
  stop,
  step,
  nSamples,
  nWindows,
  d,
  N
)
}
\arguments{
\item{attributes}{- An n x d (cell x attribute) matrix of numeric attributes for single cell data. Rownames should be cell names.}

\item{pseudotime}{- A named numeric vector of pseudotime values for cells.}

\item{randomizationParams}{- A character vector which is used to
control the production of randomized paths for comparison.}

\item{statistic}{- Allowable values are 'median', 'mean' or 'max'.}

\item{start}{- The first pseudotime value (as a percentage of the trajectory) from which to analyse the trajectory from. 
Defaults to 25\% of the way through the trajectory.}

\item{stop}{- The last pseudotime value (as a percentage of the trajectory) from which to analyse the trajectory from. 
Defaults to 75\% of the way through the trajectory.}

\item{step}{- The size of the step to take between successively later starting points in pseudotime. 
Defaults to 5\% of the trajectory length.}

\item{nSamples}{- The number of sampled paths to generate (defaults to 1000).}

\item{nWindows}{- The number of windows pseudotime should be split into to sample cells from (defaults to 10).}

\item{d}{- The dimension under consideration.  This defaults to
ncol(attributes).}

\item{N}{- The number of random paths to generated for statistical
comparison to the given path (defaults to 1000).}
}
\description{
This tests the input to analyseBranchPoint.
}
