% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chol_answers}
\alias{chol_answers}
\title{chol_answers}
\format{
A list
\describe{
Results of running analyseSingleCellTrajectory() on
a trajectory describing the
development of cholangiocytes from hepatoblasts.
}
}
\source{
Single cell data
}
\usage{
chol_answers
}
\description{
Results of running analyseSingleCellTrajectory() on
a trajectory describing the
development of cholangiocytes from hepatoblasts.
}
\keyword{datasets}
