% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chol_branch_point_results}
\alias{chol_branch_point_results}
\title{chol_branch_point_results}
\format{
A list
\describe{
Results of running analyseBranchPoint() on
a trajectory describing the
development of cholangiocytes from hepatoblasts.
}
}
\source{
Single-cell data has been obtained from GEO (GSE90047) and the script
    used for upstream processing is available at
    https://github.com/AnnaLaddach/TrajectoryGeometryData
}
\usage{
chol_branch_point_results
}
\description{
Results of running analyseBranchPoint() on
a trajectory describing the
development of cholangiocytes from hepatoblasts.
}
\keyword{datasets}
