% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chol_pseudo_time_normalised}
\alias{chol_pseudo_time_normalised}
\title{chol_pseudo_time_normalised}
\format{
A vector
\describe{
A vector of pseudotime values, normalised to range from
0 to 100, for a trajectory describing the
development of cholangiocytes from hepatoblasts.
Pseudotime values have been inferred using the SlingShot
package. The vector is named according to cell ID.
}
}
\source{
Single cell data
}
\usage{
chol_pseudo_time_normalised
}
\description{
A vector of pseudotime values, normalised to range from
0 to 100, for a trajectory describing the
development of cholangiocytes from hepatoblasts.
Pseudotime values have been inferred using the SlingShot
package. The vector is named according to cell ID.
}
\keyword{datasets}
