% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cholBranchPointResults}
\alias{cholBranchPointResults}
\title{cholBranchPointResults}
\format{
A list
\describe{
Results of running analyseBranchPoint() on 
a trajectory describing the
development of cholangiocytes from hepatoblasts.
}
}
\source{
Single cell data
}
\usage{
cholBranchPointResults
}
\description{
Results of running analyseBranchPoint() on 
a trajectory describing the
development of cholangiocytes from hepatoblasts.
}
\keyword{datasets}
