% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cholPseudoTime}
\alias{cholPseudoTime}
\title{cholPseudoTime}
\format{
A vector
\describe{
A vector of pseudotime values for a trajectory describing the
development of cholangiocytes from hepatoblasts.
Pseudotime values have been inferred using the SlingShot 
package. The vector is named according to cell ID.
}
}
\source{
Single cell data
}
\usage{
cholPseudoTime
}
\description{
A vector of pseudotime values for a trajectory describing the
development of cholangiocytes from hepatoblasts.
Pseudotime values have been inferred using the SlingShot 
package. The vector is named according to cell ID.
}
\keyword{datasets}
