% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryGeometry.R
\name{circleOnTheUnitSphere}
\alias{circleOnTheUnitSphere}
\title{Circle on the unit sphere}
\usage{
circleOnTheUnitSphere(center, radius, N = 36)
}
\arguments{
\item{center}{- The center of the circle.}

\item{radius}{- The radius of the circle.}

\item{N}{- The number of segments to approximate the circle. It
defaults to 36.}
}
\value{
This returns an approximation to the the circle as a N+1 x
    3 matrix
}
\description{
Find a circle on the unit sphere
}
\details{
Given a point on the unit sphere and a radius given
as a spherical distance, this finds the circle.

It's not clear to me this should be exported, but it's
handy to do this for testing and debugging.
}
\examples{
pole = c(1,0,0)
radius = pi / 4
circle = circleOnTheUnitSphere(pole,radius)
}
