% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{circleOnTheUnitSphereTest}
\alias{circleOnTheUnitSphereTest}
\title{This tests the inputs to circleOnTheUnitSphere}
\usage{
circleOnTheUnitSphereTest(center, radius, N)
}
\arguments{
\item{center}{- The center of the circle.}

\item{radius}{- The radius of the circle.}

\item{N}{- The number of segments to approximate the circle.}
}
\description{
This tests the inputs to circleOnTheUnitSphere
}
