% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crooked_path}
\alias{crooked_path}
\title{Crooked path}
\format{
A 14 x 3 matrix
\describe{
This path changes direction after the 5th point.
}
}
\source{
Synthetic data
}
\usage{
crooked_path
}
\description{
A path of n points in dimension d is an n x d matrix.  This
particular path is relatively crooked.
}
\keyword{datasets}
