% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crooked_path_center}
\alias{crooked_path_center}
\title{Crooked path center}
\format{
A vector of length 3
\describe{ A unit vector of length 3
    minimizing mean spherical distance to the points of the
    projection of crooked_path.
}
}
\source{
Synthetic data.
}
\usage{
crooked_path_center
}
\description{
The point on the unit sphere minimizing mean spherical distance to
    the projection of crooked_path
}
\keyword{datasets}
