% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crooked_path_projection}
\alias{crooked_path_projection}
\title{Crooked path projection}
\format{
An 8 x 3 matrix
\describe{
The projection of crooked_path[7:14,] onto the unit sphere as seen
    from crooked_path[6,] .
}
}
\source{
This was created by code in createSyntheticData.R available
    from https://github.com/AnnaLaddach/TrajectoryGeometryData
}
\usage{
crooked_path_projection
}
\description{
The projection of the last 8 points on crooked_path onto the unit
sphere as seen from the the 6th.  This is a collection of 8 unit
points in dimension 3.
}
\keyword{datasets}
