% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crooked_path_radius}
\alias{crooked_path_radius}
\title{Crooked path radius}
\format{
Numeric
\describe{
The mean spherical distance from the points of the projection of
crooked_path to the point minimizing this mean distance.
}
}
\source{
Synthetic data
}
\usage{
crooked_path_radius
}
\description{
The mean spherical distance from the points of the projection of
crooked_path to the point minimizing this mean distance.
}
\keyword{datasets}
