% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crookedPathCenter}
\alias{crookedPathCenter}
\title{Crooked path center}
\format{
A vector of length 3
\describe{ A unit vector of length 3
    minimizing mean spherical distance to the points of the
    projection of crookedPath.
}
}
\source{
Synthetic data.
}
\usage{
crookedPathCenter
}
\description{
The point on the unit sphere minimizing mean spherical distance to
    the projection so crookedPath
}
\keyword{datasets}
