% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crookedPathProjection}
\alias{crookedPathProjection}
\title{Crooked path projection}
\format{
An 8 x 3 matrix
\describe{
The projection of crookedPath[7:14,] onto the unit sphere as seen
    from crookedPath[6,] .
}
}
\source{
Synthetic data
}
\usage{
crookedPathProjection
}
\description{
The projection of the last 8 points on crookedPath onto the unit
sphere as seen from the the 6th.  This is a collection of 8 unit
points in dimension 3.
}
\keyword{datasets}
