% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crookedPathRadius}
\alias{crookedPathRadius}
\title{Crooked path radius}
\format{
Numeric
\describe{
The mean spherical distance from the points of the projection of
crookedPath to the point minimizing this mean distance.
}
}
\source{
Synthetic data
}
\usage{
crookedPathRadius
}
\description{
The mean spherical distance from the points of the projection of
crookedPath to the point minimizing this mean distance.
}
\keyword{datasets}
