% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{distanceBetweenTrajectoriesTest}
\alias{distanceBetweenTrajectoriesTest}
\title{This tests the imput to distanceBetweenTrajectories}
\usage{
distanceBetweenTrajectoriesTest(attributes1, pseudotime1, attributes2)
}
\arguments{
\item{attributes1}{- An n x d (cell x attribute) matrix of numeric attributes for the first single cell trajectory.}

\item{pseudotime1}{- A named numeric vector of pseudotime values for the first single cell trajectory, 
names should match rownames of atrributes1.}

\item{attributes2}{- An n x d (cell x attribute) matrix of numeric attributes for the sencond single cell trajectory.}
}
\description{
This tests the imput to distanceBetweenTrajectories
}
