% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryGeometry.R
\name{dummyPlaceholder}
\alias{dummyPlaceholder}
\title{A dummy function with no other purpose than to have somewhere to put
all the imports.}
\usage{
dummyPlaceholder()
}
\value{
Gornisht!
}
\description{
There really ought to be a better way to put all the importFrom comments
in one place.
}
