% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{findSphereClusterCenterTest}
\alias{findSphereClusterCenterTest}
\title{This tests the inputs to findSphereClusterCenter}
\usage{
findSphereClusterCenterTest(points, statistic, normalize)
}
\arguments{
\item{points}{- A set of n points on the (d-1) sphere given as an n
x d matrix.}

\item{statistic}{- The statistic to be minimized.  Allowable values
are 'median','mean' or 'max'.}

\item{normalize}{- If this is set to TRUE, the function will start
by normalizing the input points.}
}
\description{
This tests the inputs to findSphereClusterCenter
}
