% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{findSphericalDistanceTest}
\alias{findSphericalDistanceTest}
\title{This tests the inputs to findSphericalDistance}
\usage{
findSphericalDistanceTest(center, points, normalize)
}
\arguments{
\item{center}{- The proposed point from which distance to
the others should be measured.  This is a numerical vector
of length d.}

\item{points}{- The set of target points for which spherical
distance to the center should be calculated.  This is in the
form of a n x d matrix.}

\item{normalize}{- If this is set to TRUE, the function will start
by normalizing the input points.}
}
\value{
A successful test returns 0, otherwise execution halts.
}
\description{
This tests the inputs to findSphericalDistance
}
