% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryGeometry.R
\name{generateRandomPathsByPermutation}
\alias{generateRandomPathsByPermutation}
\title{Produce randomized paths by permutation}
\usage{
generateRandomPathsByPermutation(path, randomizationParams, N)
}
\arguments{
\item{path}{- An n x d matrix.}

\item{randomizationParams}{- A character vector used to control the
behavior of the function.}

\item{N}{- The number of paths required.}
}
\value{
This function returns a list of random paths.
}
\description{
This function produces randomized paths from a given path via
permutation of its entries.  This can be done either by random
permutation within each column thereby preserving the range of
values within each column or by random permutation of all entries
in the matrix.
}
