% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryGeometry.R
\name{generateRandomPathsBySteps}
\alias{generateRandomPathsBySteps}
\title{Produce randomized paths by steps}
\usage{
generateRandomPathsBySteps(path, randomizationParams, N)
}
\arguments{
\item{path}{- An n x d matrix.}

\item{randomizationParams}{- A character vector used to control the
behavior of the function.}

\item{N}{- The number of paths required.}
}
\value{
This function returns a list of random paths.
}
\description{
This function produces randomized paths from a given path by taking
steps in space. This can be done either requiring that these steps
have the same Euclidean length as those of the original path or
allowing them to all have unit length.  It can also be done
requiring the path to stay in the non-negative orthant or allowing
arbitrary values.
}
