% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{generateRandomPathsTest}
\alias{generateRandomPathsTest}
\title{This tests the inputs to generateRandomPaths}
\usage{
generateRandomPathsTest(path, from, to, d, randomizationParams, N)
}
\arguments{
\item{path}{- This is an mxn dimensional matrix. Each row is
considered a point.}

\item{from}{- The starting place along the path which will be
treated as the center of the sphere.}

\item{to}{- The end point of the path.}

\item{d}{- The dimension under consideration.}

\item{randomizationParams}{- A character vector controling the
randomization method used.  It's first entry must be either
'byPermutation' or 'bySteps'  See the vignette for further
details.}

\item{N}{- The number of random paths required.}
}
\description{
This tests the inputs to generateRandomPaths
}
