% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryGeometry.R
\name{generateRandomUnitVector}
\alias{generateRandomUnitVector}
\title{Generate random unit vector.}
\usage{
generateRandomUnitVector(d)
}
\arguments{
\item{d}{- The dimension.}
}
\value{
A unit vector in dimension d.
}
\description{
This function generates a random unit vector in
in dimension d.
}
