% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{generateRandomUnitVectorTest}
\alias{generateRandomUnitVectorTest}
\title{This tests the inputs for generateRandomUnitVector}
\usage{
generateRandomUnitVectorTest(d)
}
\arguments{
\item{d}{- The dimension.}
}
\value{
A successful test returns 0, otherwise execution halts.
}
\description{
This tests the inputs for generateRandomUnitVector
}
