% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{getDistanceDataForPathsTest}
\alias{getDistanceDataForPathsTest}
\title{This tests the inputs for getDistanceDataForPaths}
\usage{
getDistanceDataForPathsTest(paths, statistic)
}
\arguments{
\item{paths}{- A list of paths.  Each of these is an n x d matrix. ***Really a list?}

\item{statistic}{- Allowable values are 'median', 'mean' or 'max'.}
}
\description{
This tests the inputs for getDistanceDataForPaths
}
