% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryGeometry.R
\name{getSphericalData}
\alias{getSphericalData}
\title{This is a simplified wrapper for pathToSphericalData}
\usage{
getSphericalData(path, statistic)
}
\arguments{
\item{path}{- an m x n matrix.  Each row is considered a point}

\item{statistic}{- one of 'mean','median' or 'max'}
}
\value{
This function returns a list whose elements are the
    projections of the path to the sphere, the center for those
    projections, the median, mean or max distance from the center
    to those projections and the name of the statistic used.
}
\description{
It handles the case in which from, to and d are all
given by the dimensions of the path
}
