% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{getSphericalDataTest}
\alias{getSphericalDataTest}
\title{This tests the inputs to getSphericalData}
\usage{
getSphericalDataTest(path, statistic)
}
\arguments{
\item{path}{- an m x n matrix.  Each row is considered a point}

\item{statistic}{- one of 'mean','median' or 'max'}
}
\description{
This tests the inputs to getSphericalData
}
