% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glialAnswers}
\alias{glialAnswers}
\title{glialAnswers}
\format{A list
\describe{
Results of running analyseSingleCellTrajectory() on 
a gliogenic trajectory.
}}
\source{
Single cell data
}
\usage{
glialAnswers
}
\description{
Results of running analyseSingleCellTrajectory() on 
a gliogenic trajectory.
}
\keyword{datasets}
