% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glialPseudoTime}
\alias{glialPseudoTime}
\title{glialPseudoTime}
\format{
A vector
\describe{
A vector of pseudotime values for a gliogenic trajectory.
Pseudotime values have been inferred using the SlingShot 
package. The vector is named according to cell ID.
}
}
\source{
Single cell data
}
\usage{
glialPseudoTime
}
\description{
A vector of pseudotime values for a gliogenic trajectory.
Pseudotime values have been inferred using the SlingShot 
package. The vector is named according to cell ID.
}
\keyword{datasets}
