% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hep_attributes}
\alias{hep_attributes}
\title{hep_attributes}
\format{
A matrix
\describe{
PCA projections derived from normalised gene expression values for
single cells, and filtered for cells which feature in a trajectory
from hepatoblast to hepatocyte. The columns are the PCs and the
rows are the cells.
}
}
\source{
Single-cell data has been obtained from GEO (GSE90047) and the script
    used for upstream processing is available at
    https://github.com/AnnaLaddach/TrajectoryGeometryData
}
\usage{
hep_attributes
}
\description{
PCA projections derived from normalised gene expression values for
single cells, and filtered for cells which feature in a trajectory
from hepatoblast to hepatocyte. The columns are the PCs and the
rows are the cells.
}
\keyword{datasets}
