% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hep_pseudo_time}
\alias{hep_pseudo_time}
\title{hep_pseudo_time}
\format{
A vector
\describe{
A vector of pseudotime values for a trajectory describing the
development of hepatocytes from hepatoblasts.
Pseudotime values have been inferred using the SlingShot
package. The vector is named according to cell ID.
}
}
\source{
Single cell data
}
\usage{
hep_pseudo_time
}
\description{
A vector of pseudotime values for a trajectory describing the
development of hepatocytes from hepatoblasts.
Pseudotime values have been inferred using the SlingShot
package. The vector is named according to cell ID.
}
\keyword{datasets}
