% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hep_pseudo_time_normalised}
\alias{hep_pseudo_time_normalised}
\title{hep_pseudo_time_normalised}
\format{
A vector
\describe{
A vector of pseudotime values for a trajectory describing the
development of hepatocytes from hepatoblasts.
Pseudotime values have been inferred using the SlingShot
package. The vector is named according to cell ID.
}
}
\source{
Single-cell data has been obtained from GEO (GSE90047) and the script
    used for upstream processing is available at
    https://github.com/AnnaLaddach/TrajectoryGeometryData
}
\usage{
hep_pseudo_time_normalised
}
\description{
A vector of pseudotime values, normalised to range from
0 to 100, for a trajectory describing the
development of hepatocytes from hepatoblasts.
Pseudotime values have been inferred using the SlingShot
package. The vector is named according to cell ID.
}
\keyword{datasets}
