% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hepAnswers}
\alias{hepAnswers}
\title{hepAnswers}
\format{A list
\describe{
Results of running analyseSingleCellTrajectory() on a trajectory describing the
development of hepatocytes from hepatoblasts.
}}
\source{
Single cell data
}
\usage{
hepAnswers
}
\description{
Results of running analyseSingleCellTrajectory() on a trajectory describing the
development of hepatocytes from hepatoblasts.
}
\keyword{datasets}
