% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hepAttributes}
\alias{hepAttributes}
\title{hepAttributes}
\format{A matrix
\describe{
PCA projections derived from normalised gene expression values for 
single cells, and filtered for cells which feature in a trajectory 
from hepatoblast to hepatocyte. The columns are the PCs and the 
rows are the cells.
}}
\source{
Single cell data
}
\usage{
hepAttributes
}
\description{
PCA projections derived from normalised gene expression values for 
single cells, and filtered for cells which feature in a trajectory 
from hepatoblast to hepatocyte. The columns are the PCs and the 
rows are the cells.
}
\keyword{datasets}
