% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hepPseudoTime}
\alias{hepPseudoTime}
\title{hepPseudoTime}
\format{A vector
\describe{
A vector of pseudotime values for a trajectory describing the
development of hepatocytes from hepatoblasts.
Pseudotime values have been inferred using the SlingShot 
package. The vector is named according to cell ID.
}}
\source{
Single cell data
}
\usage{
hepPseudoTime
}
\description{
A vector of pseudotime values for a trajectory describing the
development of hepatocytes from hepatoblasts.
Pseudotime values have been inferred using the SlingShot 
package. The vector is named according to cell ID.
}
\keyword{datasets}
