% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neuralAnswers}
\alias{neuralAnswers}
\title{neuralAnswers}
\format{A list
\describe{
Results of running analyseSingleCellTrajectory() on 
a neurogenic trajectory.
}}
\source{
Single cell data
}
\usage{
neuralAnswers
}
\description{
Results of running analyseSingleCellTrajectory() on 
a neurogenic trajectory.
}
\keyword{datasets}
