% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neuralPseudoTime}
\alias{neuralPseudoTime}
\title{neuralPseudoTime}
\format{
A vector
\describe{
A vector of pseudotime values for a neurogenic trajectory.
Pseudotime values have been inferred using the SlingShot 
package. The vector is named according to cell ID.
}
}
\source{
Single cell data
}
\usage{
neuralPseudoTime
}
\description{
A vector of pseudotime values for a neurogenic trajectory.
Pseudotime values have been inferred using the SlingShot 
package. The vector is named according to cell ID.
}
\keyword{datasets}
