% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryGeometry.R
\name{orthonormalBasis}
\alias{orthonormalBasis}
\title{Find an orthonormal basis in dimension 3}
\usage{
orthonormalBasis(x)
}
\arguments{
\item{x}{- A vector of length 3}
}
\value{
This function returns an orthonormal basis in
the the form of a 3 x 3 matrix in which the first vector is
parallel to v
}
\description{
Given a vector, this normalizes it and then uses it as
the first basis vector in an orthonormal basis.  We'll use
this to find circles around points on the sphere.
}
