% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{orthoNormalBasisTest}
\alias{orthoNormalBasisTest}
\title{This tests the input to orthoNormalBasis}
\usage{
orthoNormalBasisTest(x)
}
\arguments{
\item{x}{- A numerical vector of length 3}
}
\description{
This tests the input to orthoNormalBasis
}
