% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{pathProgressionTest}
\alias{pathProgressionTest}
\title{This tests the inputs for pathProgression}
\usage{
pathProgressionTest(path, from, to, d, direction)
}
\arguments{
\item{path}{- An n x d matrix}

\item{from}{- The point along the path to be taken as the starting
point.}

\item{to}{- The point along the path to be used as the end point.}

\item{d}{- The dimension to be used.}

\item{direction}{- A non-zero numeric whose length is the the
dimension.}
}
\value{
A successful test returns 0, otherwise execution halts.
}
\description{
This tests the inputs for pathProgression
}
