% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{pathToSphericalDataTest}
\alias{pathToSphericalDataTest}
\title{This tests the inputs to pathToSphericalData}
\usage{
pathToSphericalDataTest(path, from, to, d, statistic)
}
\arguments{
\item{path}{- This is an mxn dimensional matrix. Each row is
considered a point.}

\item{from}{- The starting place along the path which will be
treated as the center of the sphere.}

\item{to}{- The end point of the path.}

\item{d}{- The dimension under consideration.}

\item{statistic}{- One of 'median', 'mean' or 'max'}
}
\value{
A successful test returns 0, otherwise execution halts.
}
\description{
This tests the inputs to pathToSphericalData
}
