% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{plotPathProjectionCenterAndCircleTest}
\alias{plotPathProjectionCenterAndCircleTest}
\title{This tests the inputs to plotPathProjectionCenterAndCircle}
\usage{
plotPathProjectionCenterAndCircleTest(
  path,
  from,
  to,
  projection,
  center,
  radius,
  color,
  circleColor,
  pathPointSize,
  projectionPointSize,
  scale,
  newFigure
)
}
\arguments{
\item{path}{- A path of dimension 3 in the form of an N x 3 matrix.}

\item{from}{- The starting place of the section under
consideration.  This is used for marking the relevant
portion.}

\item{to}{- Likewise.}

\item{projection}{- The projection of the relevant portion of the
path.}

\item{center}{- The center of the projection points.}

\item{radius}{- The radius of the circle.}

\item{color}{- The color to use for this path and its associated
data.}

\item{circleColor}{- Sets the colour of the circle.}

\item{pathPointSize}{- Sets the size of points which represent the 
path.}

\item{projectionPointSize}{- Sets the size of points which represent the 
projected path.}

\item{scale}{- The path will be start (its actual start) at 0 and
will be scaled so that its most distant point will be at this
distance from the origin.  This is to keep it comparable in
size to the sphere.}

\item{newFigure}{- When plotting a single figure or the first of
multiple figures, this should be set to TRUE which is its
default.  Otherwise, set this to FALSE in order to add
additional paths to the same figure.}
}
\value{
A successful test returns 0, otherwise execution halts.
}
\description{
This tests the inputs to plotPathProjectionCenterAndCircle
}
