% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{projectPathToSphereTest}
\alias{projectPathToSphereTest}
\title{This tests the inputs to projectPathToSphere}
\usage{
projectPathToSphereTest(path, from, to, d)
}
\arguments{
\item{path}{- This is an mxn dimensional matrix. Each row is
considered a point.}

\item{from}{- The starting place along the path which will be
treated as the center of the sphere.}

\item{to}{- The end point of the path.}

\item{d}{- The dimension under consideration.}
}
\description{
This tests the inputs to projectPathToSphere
}
