% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{samplePathTest}
\alias{samplePathTest}
\title{This tests the inputs for samplePath.}
\usage{
samplePathTest(attributes, pseudotime, nWindows)
}
\arguments{
\item{attributes}{- An n x d (cell x attribute) matrix of numeric attributes for single cell data. Rownames should be cell names.}

\item{pseudotime}{- A named numeric vector of pseudotime values for cells.}

\item{nWindows}{- The number of windows pseudotime should be split into to sample cells from. Defaults to 10.}
}
\description{
This tests the inputs for samplePath.
}
