% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_cell_matrix}
\alias{single_cell_matrix}
\title{single_cell_matrix}
\format{
A matrix
\describe{
PCA projections derived from normalised gene expression values for
single cells. The columns are the PCs and the rows are the cells.
}
}
\source{
Single cell data
}
\usage{
single_cell_matrix
}
\description{
PCA projections derived from normalised gene expression values for
single cells. The columns are the PCs and the rows are the cells.
}
\keyword{datasets}
