% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{singleCellMatrix}
\alias{singleCellMatrix}
\title{singleCellMatrix}
\format{
A matrix
\describe{
PCA projections derived from normalised gene expression values for 
single cells. The columns are the PCs and the rows are the cells.
}
}
\source{
Single cell data
}
\usage{
singleCellMatrix
}
\description{
PCA projections derived from normalised gene expression values for 
single cells. The columns are the PCs and the rows are the cells.
}
\keyword{datasets}
