% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{straight_path}
\alias{straight_path}
\title{Straight path}
\format{
A 14 x 3 matrix
\describe{
The path is roughly in the (1,0,0) direction.
}
}
\source{
This was created by code in createSyntheticData.R available
    from https://github.com/AnnaLaddach/TrajectoryGeometryData
}
\usage{
straight_path
}
\description{
A path of n points in dimension d is an n x d matrix.  This
particular path is relatively straight.
}
\keyword{datasets}
