% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{straight_path_center}
\alias{straight_path_center}
\title{Straight path center}
\format{
A vector of length 3
\describe{
A unit vector of length 3 minimizing distance to the points of the
    projection of straight_path.
}
}
\source{
This was created by code in createSyntheticData.R available
    from https://github.com/AnnaLaddach/TrajectoryGeometryData
}
\usage{
straight_path_center
}
\description{
The point on the unit sphere minimizing mean spherical distance to
    the projection so straight_path
}
\keyword{datasets}
