% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{straight_path_projection}
\alias{straight_path_projection}
\title{Straight path projection}
\format{
A 13 x 3 matrix
\describe{
The projection of straight_path[2:14,] onto the unit sphere as seen
    from straight_path[1,] .
}
}
\source{
This was created by code in createSyntheticData.R available
    from https://github.com/AnnaLaddach/TrajectoryGeometryData
}
\usage{
straight_path_projection
}
\description{
The projection of straight_path onto the unit sphere.  This is a
collection of 13 unit points in dimension 3
}
\keyword{datasets}
