% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{straightPath}
\alias{straightPath}
\title{Straight path}
\format{
A 14 x 3 matrix
\describe{
The path is roughly in the (1,0,0) direction.
}
}
\source{
Synthetic data
}
\usage{
straightPath
}
\description{
A path of n points in dimension d is an n x d matrix.  This
particular path is relatively straight.
}
\keyword{datasets}
