% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{straightPathCenter}
\alias{straightPathCenter}
\title{Straight path center}
\format{A vector of length 3
\describe{
A unit vector of length 3 minimizing distance to the points of the
    projection of straightPath.
}}
\source{
Synthetic data.
}
\usage{
straightPathCenter
}
\description{
The point on the unit sphere minimizing mean spherical distance to
    the projection so straightPath
}
\keyword{datasets}
