% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{straightPathProjection}
\alias{straightPathProjection}
\title{Straight path projection}
\format{
A 13 x 3 matrix
\describe{
The projection of straightPath[2:14,] onto the unit sphere as seen
    from straightPath[1,] .
}
}
\source{
Synthetic data
}
\usage{
straightPathProjection
}
\description{
The projection of straightPath onto the unit sphere.  This is a
collection of 13 unit points in dimension 3
}
\keyword{datasets}
