% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{straightPathRadius}
\alias{straightPathRadius}
\title{Straight path radius}
\format{Numeric
\describe{
The mean spherical distance from the points of the projection of
straightPath to the point minimizing this mean distance.
}}
\source{
Synthetic data
}
\usage{
straightPathRadius
}
\description{
The mean spherical distance from the points of the projection of
straightPath to the point minimizing this mean distance.
}
\keyword{datasets}
