% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryGeometry.R
\name{testPathForDirectionality}
\alias{testPathForDirectionality}
\title{Test a path for directionality}
\usage{
testPathForDirectionality(
  path,
  from = 1,
  to = nrow(path),
  d = ncol(path),
  statistic,
  preserveLengths = TRUE,
  N
)
}
\arguments{
\item{path}{- An n x m matrix representing a series of n points in
dimension m.}

\item{from}{- The starting place along the path which will be
treated as the center of the sphere.  This defaults to 1.}

\item{to}{- The end point of the path.  This defaults to
nrow(path).}

\item{d}{- The dimension under consideration.  This defaults to
ncol(path)}

\item{statistic}{- Allowable values are 'median', 'mean' or 'max'}

\item{preserveLengths}{- If this is set to TRUE, the length of the
steps in the randomized paths will match those of the given
path.}

\item{N}{- The number of random paths to generated for statistical
comparison to the given path.}
}
\value{
This returns a p-value for the directionality of the given
    path.
}
\description{
This is the core function of this package.  It takes a path, and a
choice of statistical measure and computes a statistical significance
for the directionality of that path.
}
