% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{testPathForDirectionalityTest}
\alias{testPathForDirectionalityTest}
\title{Test the inputs to testPathForDirectionality}
\usage{
testPathForDirectionalityTest(
path,
from,
to,
d,
randomizationParams,
statistic,
N
)
}
\arguments{
\item{path}{- An n x m matrix representing a series of n points in
dimension m.}

\item{from}{- The starting place along the path which will be
treated as the center of the sphere.}

\item{to}{- The end point of the path.}

\item{d}{- The dimension under consideration.}

\item{randomizationParams}{- A character vector which is used to
control the production of randomized paths for comparison.}

\item{statistic}{- Allowable values are 'median', 'mean' or 'max'}

\item{N}{- The number of random paths to generated for statistical
comparison to the given path.}
}
\value{
A successful test returns 0, otherwise execution halts.
}
\description{
Test the inputs to testPathForDirectionality
}
