% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryGeometry.R
\name{varianceInDirection}
\alias{varianceInDirection}
\title{Portion of cell variance in a given direction}
\usage{
varianceInDirection(direction, cells)
}
\arguments{
\item{direction}{- a numeric giving the direction expressed in PCs}

\item{cells}{- a set of cells given as a matrix where each row is a
cell and each column is a PC.  This must have at least as many
PCs as the direction}
}
\value{
The fraction of total variance (across that many PCs) in
    the given direction.
}
\description{
This function takes a direction and a set of cells epressed in PCs
and determines the fraction of total variance in the given
direction.
}
