% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{visualiseBranchPointStatsTest}
\alias{visualiseBranchPointStatsTest}
\title{This tests the input to visualiseBranchPointStats}
\usage{
visualiseBranchPointStatsTest(branchPointData, average)
}
\arguments{
\item{branchPointData}{- the result of analyseBranchPoint}

\item{average}{- if there are multiple distances available for each}
}
\value{
A successful test returns 0, otherwise execution halts.
}
\description{
This tests the input to visualiseBranchPointStats
}
