% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryGeometry.R
\name{visualiseTrajectoryStats}
\alias{visualiseTrajectoryStats}
\title{Visualise Trajectory Stats}
\usage{
visualiseTrajectoryStats(
  answers,
  metric,
  distanceMetric = "Mean",
  answers2 = list()
)
}
\arguments{
\item{answers}{- the result of analyseSingleCellTrajectory}

\item{metric}{- either pvalue or distance}

\item{distanceMetric}{- if there are multiple distances available for each 
sampled trajectory choose to plot the mean or median distance.}

\item{answers2}{either an empty list or the result of analyseSingleCellTrajectory}
}
\value{
a list containing:
 stats - output of wilcox test
 values - dataframe containing plotted data in long format
 plot - ggplot object
}
\description{
This function creates boxplots for comparisons of metrics for sampled
paths to random paths. It can also create plots for comparing two sets of
sampled paths by providing the answers2 argument.
}
