% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTestFunctions.R
\name{visualiseTrajectoryStatsTest}
\alias{visualiseTrajectoryStatsTest}
\title{This tests the input to visualiseTrajectoryStats}
\usage{
visualiseTrajectoryStatsTest(traj1Data, metric, average, traj2Data)
}
\arguments{
\item{traj1Data}{- the result of analyseSingleCellTrajectory}

\item{metric}{- either "pValue" or "distance"}

\item{average}{- if there are multiple distances available for each 
sampled trajectory, calculate the average using mean or median (defaults to mean).}

\item{traj2Data}{either an empty list or the result of analyseSingleCellTrajectory}
}
\value{
A successful test returns 0, otherwise execution halts.
}
\description{
This tests the input to visualiseTrajectoryStats
}
