% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryGeometry.R
\name{generateRandomPaths}
\alias{generateRandomPaths}
\title{Produce random paths modeled on a given path}
\usage{
generateRandomPaths(
  path,
  start = 1,
  end = nrow(path),
  d = ncol(path),
  preserveLengths = TRUE,
  N
)
}
\arguments{
\item{path}{- This is an mxn dimensional matrix. Each row is
considered a point.}

\item{start}{- The starting place along the path which will be
treated as the center of the sphere.  This defaults to 1.}

\item{end}{- The end point of the path.  This defaults to
nrow(path).}

\item{d}{- The dimension under consideration.  This defaults to
ncol(path)}

\item{preserveLengths}{- If this is set to TRUE, the steps of the
random paths have the same euclidean length as those of the
give path.  Otherwise they hav length 1.}

\item{N}{- The number of random paths required.}
}
\value{
This function returns a list of random paths.  Each path is
    a matrix.  Note that each random path begins at the origin.
}
\description{
This function takes a path and produces N random paths of the same
dimension and length based on it. The steps along the random paths
are based on uniform sampling of the unit sphere.  If 
preserveLengths is set to TRUE, the random paths of steps of the
same length as the original path.  Otherwise, the steps are all
length 1.
}
