% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oscillation}
\alias{oscillation}
\title{Oscillation}
\format{
A matrix
\describe{
This a path which prepends small oscillations to straight path.
Its purpose is to illustrate instability of spherical projection
near the beginning of a path.
}
}
\source{
This was created by code in createSyntheticData.R available from
https://github.com/AnnaLaddach/TrajectoryGeometryData
}
\usage{
oscillation
}
\description{
This a path which prepends small oscillations to straight path.
Its purpose is to illustrate instability of spherical projection
near the beginning of a path.
}
\keyword{datasets}
