% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{straight_path_radius}
\alias{straight_path_radius}
\title{Straight path radius}
\format{
Numeric
\describe{
The mean spherical distance from the points of the projection of
straight_path to the point minimizing this mean distance.
}
}
\source{
Synthetic data
}
\usage{
straight_path_radius
}
\description{
The mean spherical distance from the points of the projection of
straight_path to the point minimizing this mean distance.
}
\keyword{datasets}
